Playing around with the Image/Video Thumbnail stuff.
I might make it so the user gets sent to YouTube if they are actually on a mobile device. I also would like to make it a bit more responsive for the Modal Size. I still might abandon it and go with a standard library, I just haven't found one I like yet.